refactor(release-notes): extract pure changelog/store logic into testable modules
Move changelog parsing + scaffold building into release-notes-changelog.ts and the EAS store.config merge into release-notes-store.ts. Both are free of import.meta and fs so jest (babel-preset-expo / Hermes) can import them directly; the CLI scripts keep the import.meta.dir bootstrap and pass APP_ROOT in (resolveVersion now takes appRoot).
